+++ test/UndisclosedPasswordTest.php
@@ -196,6 +196,15 @@ class UndisclosedPasswordTest extends TestCase
     }
 
     /**
+     * Test that the message templates are getting initialized via
+     * the parent::_construct call
+     */
+    public function testMessageTemplatesAreInitialized() : void
+    {
+        $this->assertNotEmpty($this->validator->getMessageTemplates());
+    }
+
+    /**
      * Testing that we capture any failures when trying to connect with
      * the HIBP web service.
      *