+++ CHANGELOG.md
@@ -24,6 +24,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 - [#57](https://github.com/laminas/laminas-validator/pull/57) removes redundant third argument in `UndisclosedPassword` validator constructor.
 
+- [#53](https://github.com/laminas/laminas-validator/pull/53) fixes `UndisclosedPassword` validator to call parent constructor on instantiation. 
+
 ## 2.13.1 - 2020-01-15
 
 ### Added